+++ src/main/java/lotto/domain/Merchant.java
@@ -1,7 +1,5 @@
 package lotto.domain;
 
-import lotto.Lotto;
-
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
